+++ src/go/rpk/pkg/cli/cmd/acl/list.go
@@ -22,7 +22,7 @@ import (
 	"github.com/twmb/types"
 )
 
-func NewListCommand(fs afero.Fs) *cobra.Command {
+func newListCommand(fs afero.Fs) *cobra.Command {
 	var a acls
 	var printAllFilters bool
 	cmd := &cobra.Command{
